Refactoring Cloud SQL dao and cloud sql migration Dao
