+++ Paco-Server/src/com/google/sampling/experiential/dao/CSEventDao.java
@@ -1,0 +1,14 @@
+package com.google.sampling.experiential.dao;
+
+import java.sql.SQLException;
+import java.text.ParseException;
+
+import com.google.sampling.experiential.model.Event;
+
+public interface CSEventDao {
+
+  boolean insertSingleEventOnly(Event event) throws SQLException, ParseException;
+
+  boolean insertSingleEventOnlyWithExperimentInfo(Event event) throws SQLException, ParseException;
+
+}
