+++ Paco-Server/src/com/google/sampling/experiential/dao/CSFailedEventDao.java
@@ -1,0 +1,14 @@
+package com.google.sampling.experiential.dao;
+
+import java.sql.SQLException;
+import java.util.Map;
+
+public interface CSFailedEventDao {
+
+  boolean insertFailedEvent(String failedJson, String reason, String comments);
+
+  Map<Long, String> getFailedEvents() throws SQLException;
+
+  boolean updateFailedEventsRetry(Long id, String reprocessed) throws SQLException;
+
+}
