+++ Paco-Server/src/com/google/sampling/experiential/server/migration/MigrationBackendServlet.java
@@ -16,9 +16,8 @@
  */
 package com.google.sampling.experiential.server.migration;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.PrintStream;
+import java.util.Map;
 import java.util.logging.Logger;
 
 import javax.servlet.ServletException;
@@ -32,6 +31,7 @@ import org.joda.time.format.DateTimeFormat;
 
 import com.google.appengine.api.ThreadManager;
 import com.google.appengine.api.users.UserService;
+import com.google.sampling.experiential.server.ExceptionUtil;
 import com.google.sampling.experiential.server.HttpUtil;
 import com.google.sampling.experiential.server.ReportJobStatusManager;
 import com.pacoapp.paco.shared.util.Constants;
@@ -45,14 +45,12 @@ import com.pacoapp.paco.shared.util.Constants;
 public class MigrationBackendServlet extends HttpServlet {
 
   public static final Logger log = Logger.getLogger(MigrationBackendServlet.class.getName());
-  private UserService userService;
 
   @Override
   protected void doGet(final HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
     log.info("MIGRATE BACKEND");
     final String requestorEmail = getRequestorEmail(req);
     final String migrationJobName = HttpUtil.getParam(req, "migrationName");
-    final String useTaskQueue = HttpUtil.getParam(req, "queue");
     final String cursor = HttpUtil.getParam(req, "cursor");
     final String sTime = HttpUtil.getParam(req, "startTime");
     final String eTime = HttpUtil.getParam(req, "endTime");
@@ -86,7 +84,7 @@ public class MigrationBackendServlet extends HttpServlet {
             statusMgr.failReport(requestorEmail, jobId, "Check server logs for stacktrace");
           }
         } catch (Throwable e) {
-          final String fullStack = getStackTraceAsString(e);
+          final String fullStack = ExceptionUtil.getStackTraceAsString(e);
           final String string = fullStack.length()>700 ? fullStack.substring(0, 700): fullStack;
           statusMgr.failReport(requestorEmail, jobId, e.getClass() + "." + e.getMessage() +"\n" + string);
           log.severe("Could not run migration job: " + e.getMessage());
@@ -118,12 +116,4 @@ public class MigrationBackendServlet extends HttpServlet {
     return whoParam.toLowerCase();
   }
 
-
-  public String getStackTraceAsString(Throwable e) {
-    final ByteArrayOutputStream out = new ByteArrayOutputStream();
-    PrintStream pw = new PrintStream(out);
-    e.printStackTrace(pw);
-    final String string = out.toString();
-    return string;
-  }
 }
